requirements.txt
python-dotenv
scikit-learn
transformers
sentence-transformers
torch
jsonschema
rich
sentencepiece
accelerate
langchain
tiktoken
auto-gptq
langchain-core
langchain_community
optimum
faiss-cpu
hf_xet
llama-cpp-python
langchain-huggingface